add shadow blend between dynamic and vxsm

+++ com.unity.render-pipelines.lightweight/ShaderLibrary/Shadows.hlsl
@@ -38,7 +38,7 @@ half4       _MainLightShadowOffset0;
 half4       _MainLightShadowOffset1;
 half4       _MainLightShadowOffset2;
 half4       _MainLightShadowOffset3;
-half4       _MainLightShadowData;    // (x: shadowStrength)
+float4      _MainLightShadowData;    // (x: shadowStrength)
 float4      _MainLightShadowmapSize; // (xy: 1/width and 1/height, zw: width and height)
 CBUFFER_END
